+++ resources/views/components/fields/rj-display.blade.php
@@ -10,4 +10,5 @@
             {{ $id }} - {{ $workName }}
         </strong>
     </td>
+    <td class="form-table-cell form-table-cell--long-spacer" aria-hidden="true"></td>
 </tr>